Move filter to seperate filter component and emit event when filter is changed
