+++ src/main/scripts/src/components/views/query/filter.component.html
@@ -1,0 +1,8 @@
+<div>
+  <tag-input [tags]="items"
+             [placeholder]="placeholder"
+             [autocompleteOptions]="autocompleteOptions"
+             [options]="tagOptions"
+             (onDelete)="deleteTag($event)"
+             (onAdd)="addItem($event)"></tag-input>
+</div>
