+++ src/main/scripts/src/components/views/query/query.component.html
@@ -1,8 +1,3 @@
 <div>
-  <tag-input [tags]="items"
-             [placeholder]="placeholder"
-             [autocompleteOptions]="autocompleteOptions"
-             [options]="tagOptions"
-             (onDelete)="deleteTag($event)"
-             (onAdd)="addItem($event)"></tag-input>
+  <query-filter (filterChanged)="onFilterChanged($event)"></query-filter>
 </div>
